[RW-5058][risk=no] Move DataSetService into dataset package (#3722)
* moving DataSetService from dao to dataset package

* moving try catch to service
